requirements.txt
colormath==3.0.0
colorthief==0.2.1
Flask==2.0.1
numpy==1.19.5
pandas==1.3.2
Pillow==8.4.0
requests==2.26.0
keras==2.6.*
tensorflow==2.6.0
